Initial pass at AWS Auditor
This is a script to check for AWS compliance and alert to an SNS topic.
It can be run via Lambda or alternatively AWS Config, where it will
report back to Config. Heavily influenced by
https://github.com/awslabs/aws-security-benchmark , which will be run
separately. Currently only checks for some S3 bucket compliance, but
more will be added. This will run daily via a scheduled AWS Lambda
invocation configured with Terraform.